+++ src/composites/LinePlot/hooks/use-chart-zoom.ts
