+++ src/pages/Landing/Landing.tsx
@@ -1,4 +1,3 @@
-import { PageHeader } from "../../views/PageHeader";
 import { Box, Button, Flex, Text } from "@radix-ui/themes";
 import { InitialProjectSetupDialog } from "./InitialProjectSetupDialog.tsx";
 import { useState } from "react";
