+++ src/pages/ProjectDetailsView/ProjectPanel.tsx
